+++ R/autoref.r
@@ -2,12 +2,23 @@
 #'
 #' @param schema a \code{\link{relation_schema}} object, as given by
 #'   \code{\link{synthesise}}.
+#' @inheritParams autodb
 #'
 #' @return A \code{\link{database_schema}} object, containing the given relation
 #'   schemas and the created foreign key references.
 #' @export
-autoref <- function(schema) {
-  references <- calculate_references(schema)
+#' @examples
+#' rs <- relation_schema(
+#'   list(
+#'     a_b_c = list(c("a", "b", "c", "d"), list(c("a", "b", "c"))),
+#'     a_b = list(c("a", "b", "d"), list(c("a", "b"), c("b", "d")))
+#'   ),
+#'   letters[1:4]
+#' )
+#' autoref(rs, single_ref = FALSE)
+#' autoref(rs, single_ref = TRUE)
+autoref <- function(schema, single_ref = FALSE) {
+  references <- calculate_references(schema, single_ref = single_ref)
   references <- Map(
     \(child, parent, attr) list(
       child,
@@ -22,7 +33,7 @@ autoref <- function(schema) {
   database_schema(schema, references)
 }
 
-calculate_references <- function(schema) {
+calculate_references <- function(schema, single_ref = FALSE) {
   keys <- keys(schema)
   attrs <- attrs(schema)
   # find all links for indexes (should be any candidate key instead)
@@ -38,7 +49,8 @@ calculate_references <- function(schema) {
           child_ref_attrs <- c(child_ref_attrs, child)
           parent_ref_attrs <- c(parent_ref_attrs, parent)
           ref_attrs <- c(ref_attrs, list(parent_key))
-          break
+          if (single_ref)
+            break
         }
       }
     }
@@ -49,24 +61,28 @@ calculate_references <- function(schema) {
   # we do this by abusing the remove_extraneous_dependencies function
   # for functional dependencies
   fds <- functional_dependency(
-    Map(list, names(schema)[child_ref_attrs], names(schema)[parent_ref_attrs]),
-    names(schema)
+    Map(list, names(schema)[child_ref_attrs], names(schema)[parent_ref_attrs]) |>
+      unname(),
+    names(schema),
+    unique = FALSE
   )
-  filtered_fds <- remove_extraneous_dependencies(fds)
+  unique_fds <- unique(fds)
+  unique_fd_indices <- match(fds, unique_fds)
+  filtered_fds <- remove_extraneous_dependencies(unique_fds)
   filtered_vecs <- list(
     determinant_sets = detset(filtered_fds),
     dependants = dependant(filtered_fds)
   )
 
-  relation_pairs <- fds
-  filtered_relation_pairs <- filtered_fds
-  kept <- match(filtered_relation_pairs, relation_pairs)
-  stopifnot(!anyNA(kept))
-  filtered_attrs <- ref_attrs[kept]
+  kept <- is.element(unique_fds, filtered_fds)
+  fds_kept <- kept[unique_fd_indices]
+  kept_indices <- match(fds[fds_kept], filtered_fds)
+  stopifnot(!anyNA(fds_kept))
+  filtered_attrs <- ref_attrs[fds_kept]
 
   list(
-    child = filtered_vecs$determinant_sets,
-    parent = filtered_vecs$dependants,
+    child = filtered_vecs$determinant_sets[kept_indices],
+    parent = filtered_vecs$dependants[kept_indices],
     attr = filtered_attrs
   )
 }